Add files via upload
A fun project where I wrote a simple program that encoded any file at the binary level using an Aphine Cipher.  The cipher is successful and can be decrypted.  However, Aphine ciphers are not very secure as seen by the still encrypted but recognizable disco donut.  I had a lot of fun with the math.